Fixed retain error in Parser.m
The TokenStream instance variable "input" was not retained, when
the input was set using the Parser.initWithTokenStream:
This could cause the dealloc method to release an already destroyed
input instance. Also, the dealloc method used the setInput method
to release the input. This doesn't work in the dealloc method, as
the self method is already too far gone at this point.